Put actual content on the home page
Now it's no longer just default/placeholder text from the site
generator. It's also formatted all pretty-like. My self-bio is a bit
long compared to the rest of the content, but once there's more other
content on the rest of the page, I hope it won't feel so top-heavy.

Unfortunately, the venue and time isn't quite locked down yet, so that's
just a placeholder. Similarly, the Meetup page hasn't been set up
either, so that just links to the Meetup homepage.

And, finally, the DNS still needs to be set up, so crediting Hover is a
bit premature, but it's pretty much who I'm planning to go with for the
time being.