Add a "Standards of Behavior" document
This is heavily influenced by the "Code of Conduct" template from the
Geek Feminism wiki, but heavily changed both for my own personal
stylings/feelings, and for the fact that this is a monthly user group,
not a full-fledged conference with a staff and/or entry fees. I will
likely ask a few friends to look over it, as while I hope I didn't miss
anything critical, an extra pair of eyes never hurt anybody.