Rename "Standards of Behavior" > "Code of Conduct"
I was trying to strike my own path on this topic, but to be honest, the
more I looked at "Standards of Behavior", the more I felt like it was
pre-emptively tut-tutting people, instead of simply describing how the
user group is meant to be an open and welcoming space to everyone, and
what steps are available for anyone who doesn't feel like we're living
up to the expectations (but not, by default, assuming someone will
violate the ideas).

In some cases, people feel a "Code of Conduct" is similarly
pre-emptively assuming guilty behaviors, but this certainly extends
beyond the simple naming of the idea, and as the alternate name I chose
simply exacerbates this, I've reverted to the more common name for the
idea.

I may update this again later, but for now, this makes me feel a little
bit better about how it is worded.