Add additional external links
The ICCF Holland is the charity to which proceeds from donations for Vim
development actually goes, and in the spirit of supporting and
encouraging Vim usage, it only makes sense that the group would do what
it can to support and encourage the same humanitarian causes that Vim
does.

And, since the VLP is graciously hosting us, it makes sense to give them
top billing on the links, as opposed to just embedded somewhere inside
the page.