Excise mentions of my Twitter account
While I feel there is great value in the way I use and interact with
Twitter, currently (and in general), it is saturated with lots of news
that, while important to me personally, is completely unrelated to the
fostering of knowledge in Vim.

To that end, and to keep people from potentially being turned away from
the group by the passion and fervor of my frequent and unrelated tweets,
I've created a new Twitter account where users can interact with the
group: @baltimorevim. That account is now the only Twitter account that
is mentioned on the site.

Soon, I'll see if I can find a way to easily embed that account's tweets
in the page, as they may alternate between general-purpose useful Vim
news, and items relevant to the group as a whole, but in any case,
should always be a handy addition to the information that's presented.