Add a basic README
This doesn't say much, because the majority of what we want to say is on
the site itself, but GitHub kept nagging for a README, so here it is.

I don't have my usual "License" section in here, instead, I've got a
"Copyright" section. I'll have to investigate; I think, as far as this
site, my usual MIT leanings might not be appropriate, a Creative Commons
seems like it'd work with things that are largely text content instead
of code.