Update main page with links for July Meetup
I missed updating the page for June's Meetup, but won't be so lax with
July!

There are RSVP links now for both the technical and social hours, and
though there is/are no set talk(s) yet, there are links to make it
easier on someone to try and fill in the space.