Update the last video with a slide-visible version
And, now that all our talks have slides visible, we don't need the
disclaimer about working towards that goal, either. However, there *is*
a new one, in that producing these videos requires some level of video
editing and a full copy of the slides as shown by the presenter, so that
it's not just assumed as a given because we've been able to do it so
far.