Don't display titles on the videos
Because I created a custom image to display at the start of all the
videos, there's no reason to display the titles for all of them prior to
playing; in fact, doing so cuts into the new custom title card and kinda
makes it not-so-pretty. This layout works a bit better.