Add suggestion form and two suggested talks
The talk form has been something I've been intending to do since
launching the site, and now it exists! It's not as long and
comprehensive as the "Submit a Talk" form, but it's not meant to be;
people requesting talks obviously have less of an idea about what the
talk will entail than someone planning to give a talk.

The other two ideas have also been rattling around in my own head for a
while, so here's hoping someone has the ability to give them!