Move organizer info to the FAQ page
The index page was getting a little bit long for the "quick intro" feel
I was going for; most people don't care *that* much about who *I* am,
they just want information about the group. And, while it was at the
bottom of the page, some people might see the lengthy scrollbar when
they load up the site and immediately bail instead of trying to read
everything.

But, that being said, maybe it *is* a question someone might have, so
now it's on the question page. I think this is a fair trade-off.