Move general schedule to the bottom of the index
Generally speaking, with the RSVP buttons and other links readily
available, this should clarify where and when individual meetups are
happening, which is generally what I think someone might come to this
page for. So, in an effort to give someone the most valuable information
up-front, I've rearranged it so that the "Next Meetup" section is right
after the initial intro.

It's a good thing.