Add video links/descriptions for July 2015
While the video for my own talk has yet to be edited, the raw footage
has been uploaded. Meanwhile, Tushar Dave's video, being a lightning
talk (ie, shorter), was easier to edit together, and so the final
version of that video has been uploaded and is embedded in this page.