Make the "Give This Talk!" links more intelligent
They now link to the "Submit a Talk" page in a way that allows the page
to intelligently construct a reference to a partially pre-filled Google
Docs form, so they don't have to re-type anything they saw on the
previous page. Of course, they're free to modify anything, but this
further helps kick-start the process for submitting a talk.