Add August's Meetup videos
This does not include an edited version of Tushar Dave's talk;
fortunately, the video was clear enough that his slides are generally
visible, so I am not as stressed as I usually am about this. However,
I'd still like to get an edited version up before our next meeting on
the 18th.