Remove "official" social hour, mention beer/food
I'm going to try to bring food and drink to the events, as I'm able,
because I think food and drink will help bring out more people to the
meetups.

In the meantime, the "official" social hour is being taken down. I will
likely continue to head to Yum's, but in the rare case that I don't feel
like it and nobody else is going, I won't feel any obligation as it is
not a separately-scheduled event that someone may be planning to come to
(despite not submitting an RSVP).