Put up January 2016 schedule
This is just one talk, and I'm hoping that Robert Weisser will be able
to give his talk this month, as I think it dovetails nicely with the
current writeups. However, as I've only recently contacted him and not
yet heard back, I won't get people's hopes up in case the answer comes
back as a "no".