Add Gitter links
I'm hoping this will become a nice place for us to talk as a group when
we're not at the VLP. People have been sending me some awesome links and
snippets they've written, but I don't have a mailing list! I think this
would be a neat place to share these things around.