Add the schedule for the next meetup (May 9)!
We're back! And we have the combination to the Vim Meetup. April was
taken off because scheduling got too hairy, but as a followup to my
DCBPW presentation doing an introduction to git, I figured I'd do a song
and dance on the best git-related plugin I'm aware of: Fugitive.

The buttons and stuff are all in the usual place.