Update homepage with new venue information
The VLP is, at least temporarily, no longer able to host us, but the
Digital Harbor Foundation (where the Baltimore Perl meetups have been
happening) is, so we're making sure to redirect people to the
appropriate location, and include even a large header explaining why at
the top of the page.

I also tweaked a little bit of the language here and there, and as we've
been skewing earlier in the month, updated the "third our[sic] fourth"
to "first or second" Monday.