Trim out the Gitter page/some other extra text
This home page is getting a little busy, and as nobody seems to have
wanted to use the Gitter page, I'm removing specific mention from the
home page. I also cut out the text around it and reworked it into a
version that's part of the paragraph above, so we can remove an extra
amount of blank space.