Refactor header buttons on all pages
The buttons now take up overall less space to say roughly the same
things, which among other things lets us put a button for our new
physical host, the DHF.

Additionally, we have a link to Vim.org itself, which I'm surprised I'd
left out before. And, all external-site buttons are on the second row
instead of some being on the top row and some being on the bottom. This
appeals to my sense of order.

The "Most Recent Past Meetups" button was shortened just because it
seemed excessively long-winded, but also theoretically it gives us
another year or two of room to grow, by which point I assume I'll have
redone this site somewhat anyway.